Build with stable Rust before building with Nightly
Seems like there is an incompatibility with the `Cargo.lock` files
generated between older and newer versions of Rust (most likely the new
merge-friendly format), which is affecting CircleCI in recent PRs.

This commit should hopefully allow us to start with an old lockfile,
test our changes on stable, and then subsequently upgrade to the new
lockfile format with nightly.