Switch from lazy_static to once_cell
The `OnceCell::get_or_try_init()` method will help use construct a
common error type to use throughout our library. With `lazy_static`, we
would be stuck with a borrowed version of the resulting error instead of
an owned value, meaning we cannot `?` or store it anywhere without a
lifetime.