Fix semantics of shutdown(), remove remove_hooks() for now
The previous changes, moving `shutdown()` to the specific
`RenderDoc<V100>` impl block and the addition of `remove_hooks()` to
`RenderDoc<V141>`, won't actually work as intended in practice.

This is because the `Deref` and `DerefMut` impls are not sufficient for
either method to work properly. This would require the introduction of a
`DerefMove` trait in the standard library to work. Tracking issue:

https://github.com/rust-lang/rfcs/issues/997