Allow global synchronized access to RenderDoc
Callback functions for the RenderDoc in-application API can be invoked
from any thread, but they are not re-entrant and require
synchronization. As such, it should be safe to manually implement `Send`
here and permit a `RenderDoc<V>` to be wrapped in a `Mutex` and possibly
stored in a static `once_cell::sync::Lazy`.