Start from scratch, rewrite version representation
The whole type-safe runtime downgrading API with `.into()` is honestly
an unnecessary gimmick and should be removed. Pretty much no one would
ever want nor need to downgrade to a different version of the RenderDoc
API at runtime, much less do it in a type-safe way. The types, traits,
and macro necessary for this feature is hardly worth the maintenance
burden. Additionally, the way the current codebase abuses `Deref` and
`DerefMut` with `std::mem::transmute` (!!) to mimic a kind of
inheritance pattern is a big and unnecessary red flag.

With all that said, however, the core concept of providing a type-safe
interface to RenderDoc guaranteed using traits and generics is still
valid, and perhaps exposing a fallible way to _upgrade_ your API rather
than an infallible way to downgrade would be more useful.

As such, this commit throws out all the existing code in the `renderdoc`
crate with the intention of rewriting it to be more ergonomic and user
friendly. The reimagined version model relies on a `Minimum<V: Version>`
trait to constrain API version numbers instead of `Deref{,Mut}`.