Add RenderDoc::set_{focus_toggle,capture}_keys()
This commit reimplements both these methods as they existed before,
except instead of accepting a `T: IntoIterator` we now use a zero-cost
trait similar to `AsRef` which represents zero or more `InputButton`s.
This is so we can avoid allocating to the heap unnecessarily every time
these methods are called. We also remove the `winit` integration that
converts `VirtualKeyCode` to `InputButton` here, because it's honestly
kinda useless and adds an unnecessary maintenance burden.