Remove Below<V110> where clause for save_all_initials()
Technically, this option still exists in RenderDoc after version 1.1.0;
it's just completely ignored. It hasn't been renamed nor removed, so the
use of `Below` is unwarranted here, IMO. I think the two `#[deprecated]`
attributes on the option getter and setter are more than sufficient to
convey its current status.

In an ideal world, we could somehow toggle the `#[deprecated]` on only
when `V: Minimum<V110>`, but this is unfortunately impossible in Rust.