Add DESIGN.md to brainstorm API design ideas
The work we've done so far in this branch is better in many ways than
what we currently have in `master`, but it still seems messy and not
idiomatic. It follows the RenderDoc API too literally rather than trying
to abstract over it to be closer to idiomatic Rust, with the exception
of the `Captures` API from the previous commit, which looks very nice.